+++ forui/lib/src/widgets/slider/thumb.dart
@@ -56,8 +56,8 @@ class _ThumbState extends State<Thumb> with SingleTickerProviderStateMixin {
   void didChangeDependencies() {
     super.didChangeDependencies();
     _controller = InheritedController.of(context);
-    _controller.tooltips.remove(_key, _tooltip);
-    _key = widget.min ? FSliderTooltipsController.min : FSliderTooltipsController.max;
+    // _controller.tooltips.remove(_key, _tooltip);
+    // _key = widget.min ? FSliderTooltipsController.min : FSliderTooltipsController.max;
     _controller.tooltips.add(_key, _tooltip);
   }